+++ lint_test/number_of_parameters_test.dart
@@ -18,28 +18,3 @@ class Exclude {
     return a + b + c;
   }
 }
-
-class UserDto {
-  final String email;
-  final String firstName;
-  final String id;
-
-  const UserDto({
-    required this.email,
-    required this.firstName,
-    required this.id,
-  });
-
-  /// Excluded by method_name
-  UserDto copyWith({
-    String? email,
-    String? firstName,
-    String? id,
-  }) {
-    return UserDto(
-      email: email ?? this.email,
-      firstName: firstName ?? this.firstName,
-      id: id ?? this.id,
-    );
-  }
-}