+++ CHANGELOG.md
@@ -8,7 +8,7 @@
   - `number_of_parameters`
 - BREAKING CHANGE: Renamed `excludeNames` parameter to `exclude` for `function_lines_of_code` lint.
 - Fixed an issue with `prefer_early_retrun` for throw expression
-- Added `copyWith` to `exclude` section for `number_of_parameters` lint 
+- `number_of_parameters` lint: added `copyWith` to the default exclude list.
 
 ## 0.2.3