+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.4.0
 
 - Added `allow_with_comments` parameter for `no_empty_block` lint.
+- Added `exclude_entity` parameter for `prefer_match_feli_name` lint.
 
 ## 0.3.0