+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.4.0
 
 - Added `allow_with_comments` parameter for `no_empty_block` lint.
-- Fixed an issue with `avoid_using_api` for extensions
+- Added extension support for `avoid_using_api`
 ## 0.3.0
 
 - Added `exclude` parameter for the following lints: